+++ tests/delete.rs
@@ -1,10 +1,11 @@
-use std::error::Error;
+use core::error::Error;
 
 use similar_asserts::assert_eq;
-use wayfind::{Router, errors::DeleteError};
+use wayfind::Router;
+use wayfind::errors::DeleteError;
 
 #[test]
-fn test_delete() -> Result<(), Box<dyn Error>> {
+fn delete() -> Result<(), Box<dyn Error>> {
     let mut router = Router::new();
     router.insert("/test", 1)?;