+++ tests/insert.rs
@@ -1,13 +1,11 @@
-use std::error::Error;
+use core::error::Error;
 
 use similar_asserts::assert_eq;
-use wayfind::{
-    Router,
-    errors::{InsertError, TemplateError},
-};
+use wayfind::Router;
+use wayfind::errors::{InsertError, TemplateError};
 
 #[test]
-fn test_insert_conflict_static() -> Result<(), Box<dyn Error>> {
+fn insert_conflict_static() -> Result<(), Box<dyn Error>> {
     let mut router = Router::new();
     router.insert("/test", 1)?;
 
@@ -26,7 +24,7 @@ fn test_insert_conflict_static() -> Result<(), Box<dyn Error>> {
 }
 
 #[test]
-fn test_insert_conflict_dynamic() -> Result<(), Box<dyn Error>> {
+fn insert_conflict_dynamic() -> Result<(), Box<dyn Error>> {
     let mut router = Router::new();
     router.insert("/<id>", 1)?;
 
@@ -48,7 +46,7 @@ fn test_insert_conflict_dynamic() -> Result<(), Box<dyn Error>> {
 }
 
 #[test]
-fn test_insert_conflict_dynamic_structural() -> Result<(), Box<dyn Error>> {
+fn insert_conflict_dynamic_structural() -> Result<(), Box<dyn Error>> {
     let mut router = Router::new();
     router.insert("/<id>", 1)?;
 
@@ -70,7 +68,7 @@ fn test_insert_conflict_dynamic_structural() -> Result<(), Box<dyn Error>> {
 }
 
 #[test]
-fn test_insert_conflict_wildcard() -> Result<(), Box<dyn Error>> {
+fn insert_conflict_wildcard() -> Result<(), Box<dyn Error>> {
     let mut router = Router::new();
     router.insert("/<*catch_all>", 1)?;
 
@@ -92,7 +90,7 @@ fn test_insert_conflict_wildcard() -> Result<(), Box<dyn Error>> {
 }
 
 #[test]
-fn test_insert_conflict_wildcard_structural() -> Result<(), Box<dyn Error>> {
+fn insert_conflict_wildcard_structural() -> Result<(), Box<dyn Error>> {
     let mut router = Router::new();
     router.insert("/<*catch_all>", 1)?;
 
@@ -114,7 +112,7 @@ fn test_insert_conflict_wildcard_structural() -> Result<(), Box<dyn Error>> {
 }
 
 #[test]
-fn test_insert_conflict_end_wildcard() -> Result<(), Box<dyn Error>> {
+fn insert_conflict_end_wildcard() -> Result<(), Box<dyn Error>> {
     let mut router = Router::new();
     router.insert("/<*catch_all>", 1)?;
 
@@ -136,7 +134,7 @@ fn test_insert_conflict_end_wildcard() -> Result<(), Box<dyn Error>> {
 }
 
 #[test]
-fn test_insert_conflict_end_wildcard_structural() -> Result<(), Box<dyn Error>> {
+fn insert_conflict_end_wildcard_structural() -> Result<(), Box<dyn Error>> {
     let mut router = Router::new();
     router.insert("/<*catch_all>", 1)?;
 
@@ -158,7 +156,7 @@ fn test_insert_conflict_end_wildcard_structural() -> Result<(), Box<dyn Error>> 
 }
 
 #[test]
-fn test_insert_duplicate_parameter() {
+fn insert_duplicate_parameter() {
     let mut router = Router::new();
     let insert = router.insert("/<*id>/users/<id>", 3);
     assert_eq!(