+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-...
+### Removed
+
+- Unreachable `TemplateError::EmptyAngles` error variant has been removed.
 
 ## [0.9.0] - 2025-08-27